simplified the syncStateMachine implementation
